simple-agent-langgraph/requirements.txt
# Core
llama-stack-client==0.3.1
python-dotenv==1.2.1

# For Testing

